refactor(code_gen): extract call_* methods into calls.rs
